REMS/requirement.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
pillow==10.2.0
PyJWT==2.8.0
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
typing-extensions==4.9.0
urllib3==2.2.0

